modified datatypes and removed validations, added two routes for getting user profile, requesting all data from body

+++ app/models/candidate.education.model.js
@@ -23,7 +23,7 @@ module.exports = (sequelize, Sequelize) => {
             type: Sequelize.INTEGER
         },
         obtained_gpa: {
-            type: Sequelize.INTEGER
+            type: Sequelize.DECIMAL(10,1)
         }
     }, {
         timestamps: false,
